+++ Criatura.cpp
@@ -58,98 +58,5 @@ bool Criatura::estaViva() const {
     return energia > 0;
 }
 
-// Implementación de Herbívoro
-Herbivoro::Herbivoro(int x, int y, int energiaInicial)
-    : Criatura(x, y, energiaInicial) {
-    tipo = "Herbivoro";
-}
 
-Herbivoro::~Herbivoro() {
-}
 
-void Herbivoro::Operacion() {
-    buscarPlantas();
-    consumirEnergia(3); // Los herbívoros consumen menos energía
-}
-
-void Herbivoro::Update() {
-    incrementarEdad();
-    consumirEnergia(2); // Metabolismo base
-}
-
-Criatura* Herbivoro::reproducirse() {
-    if (puedeReproducirse()) {
-        consumirEnergia(50);
-        return new Herbivoro(posX, posY, 80);
-    }
-    return nullptr;
-}
-
-void Herbivoro::buscarPlantas() {
-    // Lógica para buscar plantas cercanas
-    cout << "Herbivoro buscando plantas..." << endl;
-}
-
-// Implementación de Carnívoro
-Carnivoro::Carnivoro(int x, int y, int energiaInicial)
-    : Criatura(x, y, energiaInicial) {
-    tipo = "Carnivoro";
-}
-
-Carnivoro::~Carnivoro() {
-}
-
-void Carnivoro::Operacion() {
-    cazarPresa();
-    consumirEnergia(8); // Los carnívoros consumen más energía
-}
-
-void Carnivoro::Update() {
-    incrementarEdad();
-    consumirEnergia(5); // Metabolismo base más alto
-}
-
-Criatura* Carnivoro::reproducirse() {
-    if (puedeReproducirse()) {
-        consumirEnergia(70);
-        return new Carnivoro(posX, posY, 120);
-    }
-    return nullptr;
-}
-
-void Carnivoro::cazarPresa() {
-    // Lógica para cazar otras criaturas
-    cout << "Carnivoro cazando presa..." << endl;
-}
-
-// Implementación de Omnívoro
-Omnivoro::Omnivoro(int x, int y, int energiaInicial)
-    : Criatura(x, y, energiaInicial) {
-    tipo = "Omnivoro";
-}
-
-Omnivoro::~Omnivoro() {
-}
-
-void Omnivoro::Operacion() {
-    buscarAlimento();
-    consumirEnergia(5); // Consumo medio de energía
-}
-
-void Omnivoro::Update() {
-    incrementarEdad();
-    consumirEnergia(3); // Metabolismo base medio
-}
-
-Criatura* Omnivoro::reproducirse() {
-    if (puedeReproducirse()) {
-        consumirEnergia(60);
-        return new Omnivoro(posX, posY, 100);
-    }
-    return nullptr;
-}
-
-void Omnivoro::buscarAlimento() {
-    // Lógica para buscar tanto plantas como presas
-    cout << "Omnivoro buscando alimento..." << endl;
-}
