+++ TomaAgua.cpp
@@ -1,0 +1,29 @@
+#include"TomaAgua.h"
+#include"Recurso.h"
+
+// ============================================================================
+// IMPLEMENTACIÓN DE TOMAAGUA
+// ============================================================================
+
+TomaAgua::TomaAgua(int rango) : Alimentacion(rango, "Agua") {
+}
+
+TomaAgua::~TomaAgua() {
+}
+
+bool TomaAgua::ejecutar(Criatura* criatura) {
+    if (criatura == nullptr) {
+        return false;
+    }
+
+    Recurso* agua = buscarRecurso(criatura);
+    if (agua != nullptr) {
+        return consumirRecurso(criatura, agua);
+    }
+
+    return false;
+}
+
+bool TomaAgua::esRecursoCompatible(Recurso* recurso) {
+    return (recurso != nullptr && recurso->getTipo() == "Agua");
+}
